Adjust the FRR container to the rhel9 changes
Replace tini which is not available in rhel9 with catatonit and add a
symlink so we won't require changes in the images.
Copy the content of /usr/libexec/frr (new location under rhel9) to
/usr/lib (where metallb expects to find the content).
Add the frr user to the frrvty group as the rpm does not do that
anymore.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>